Photo Upload for Contact Form
Using the personal EmailJS paid subscription (500 kB max limit per file - $11 per month), the file upload feature has been added to the contact form. The uploaded file will sent as an attachment to the email. The file isn't loaded to the firebase. The file type must be image type. If the client decides to not purchase the subscription, there's no support for sending the attachment along with the file.